Extract module kore.lexer
